Add unified tool registry
