+++ backend/tests/test_matrix_agents.py
@@ -3,7 +3,8 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from backend.agents.matrix_agents import SkillSelectorAgent
-from backend.skills.matrix_skills import MatrixSkill, SkillRegistry
+from backend.core.tool_registry import MatrixSkill
+from backend.skills.matrix_skills import SkillRegistry
 
 
 class MockSkill(MatrixSkill):
