+++ backend/tests/test_tool_sandbox.py
@@ -2,7 +2,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from backend.core.base_tool import BaseRaptorTool
+from backend.core.tool_registry import BaseRaptorTool
 
 
 class FailingTool(BaseRaptorTool):
